ref(seer): split components to new directory
